+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Polyverse Boost API on Node.js
 
 ## Version 0.5.1: January 2nd, 2024
 ### New Features
-- Added API /user_project/:org/:project (DELETE) to delete user project data
+- Added APIs /user_project/:org/:project goals, analysis and data_references to delete user project data (DELETE)
 - Added placeholder API /user_project/:org/:project/data_references (POST) to create/upload user data
 
 ### Enhancements